+++ src/app/locales/en/components/file-browser/fb-metadata-rdf.js
@@ -1,0 +1,3 @@
+export default {
+  invalidXml: 'XML is not valid',
+};