+++ src/app/services/metadata-manager.js
@@ -11,6 +11,8 @@ import Service, { inject as service } from '@ember/service';
 import { get } from '@ember/object';
 import gri from 'onedata-gui-websocket-client/utils/gri';
 import { entityType as fileEntityType } from 'oneprovider-gui/models/file';
+import _ from 'lodash';
+import { Promise } from 'rsvp';
 
 const BackendMetadataType = Object.freeze({
   xattrs: 'xattrs',
@@ -28,6 +30,34 @@ function metadataGri(fileId, metadataType, scope = 'private') {
   });
 }
 
+const serializers = {
+  xattrs: _.identity,
+  /**
+   * @param {String|Object} data
+   * @returns {Object}
+   */
+  json(data) {
+    if (typeof data === 'string') {
+      return JSON.parse(data);
+    } else {
+      return data;
+    }
+  },
+  rdf: _.identity,
+};
+
+const deserializers = {
+  xattrs: _.identity,
+  /**
+   * @param {String} data
+   * @returns {Object}
+   */
+  json(data) {
+    return JSON.stringify(data, null, 2);
+  },
+  rdf: _.identity,
+};
+
 export default Service.extend({
   onedataGraph: service(),
 
@@ -42,23 +72,32 @@ export default Service.extend({
       operation: 'get',
       gri: metadataGri(get(file, 'entityId'), metadataType, scope),
       subscribe: false,
-    });
+    }).then(data => deserializers[metadataType](data.metadata));
   },
 
   /**
    * @param {Models.File} file 
    * @param {String} metadataType one of: xattrs, json, rdf
-   * @param {any} metadata Object for xattrs and JSON, String for RDF
+   * @param {any} metadata Object for xattrs, String for RDF and JSON
    * @returns {Promise<Object>} with `metadata` key
    */
   setMetadata(file, metadataType, metadata) {
-    return this.get('onedataGraph').request({
-      operation: 'create',
-      gri: metadataGri(get(file, 'entityId'), metadataType),
-      data: {
-        metadata,
-      },
-      subscribe: false,
+    const onedataGraph = this.get('onedataGraph');
+    return new Promise((resolve, reject) => {
+      try {
+        resolve(serializers[metadataType](metadata));
+      } catch (error) {
+        reject(error);
+      }
+    }).then(metadata => {
+      return onedataGraph.request({
+        operation: 'create',
+        gri: metadataGri(get(file, 'entityId'), metadataType),
+        data: {
+          metadata,
+        },
+        subscribe: false,
+      });
     });
   },